requirements.txt
PySide6<6.10
Pillow>=11.3.0
psutil>=5.9.5
requests>=2.31.0
python-slugify>=1.2.1
raven>=6.0.0
packaging>=23.2
mozjpeg-lossless-optimization>=1.2.0
natsort>=8.4.0
distro>=1.8.0
numpy>=1.22.4
PyMuPDF>=1.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
tqdm>=4.66.1
PyQt6>=6.5.0
fpdf2>=2.7.4
Pillow>=9.3.0